renamed improvedAdmin to improved_admin
