+++ companies/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 
 from .models import *
 
-from improvedAdmin import ModelAdminImproved
+from improved_admin import ModelAdminImproved
 
 
 class CompanyAddressInline(admin.StackedInline):
