+++ events/admin.py
@@ -8,7 +8,7 @@ from .models import (
     SignupQuestion,
     ParticipantCheckIn,
 )
-from improvedAdmin import ModelAdminImproved
+from improved_admin import ModelAdminImproved
 
 
 @admin.register(Participant)
