+++ matching/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import User
-from improvedAdmin import ModelAdminImproved
+from improved_admin import ModelAdminImproved
 
 from .models import (
     Question,
